+++ .github/workflows/paddle_ci.yaml
@@ -36,7 +36,6 @@ jobs:
         run: |
           pip install ${{ matrix.python-version-with-wheel[1] }}
           pip install ".[paddle-test]"
-          pip install numpy==1.24.3
 
       - uses: actions/cache@v3
         with:
